requirements.txt
Django==2.1.7
djangorestframework==3.9.2
kiwisolver==1.0.1
nose==1.3.7
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
drf-yasg=1.14.0